Pin CI builds to node v16 (#138)
This is probably worth upgrading, but at the very least this makes our CI use the same (major) version as `package.json`, `.node-version`, etc dictate is our supported version.

Upgrading to v18 in the future should be easy, just change the version in this file.  If we want to support both, we can do the strategy-matrix thing that the node template shows.  But since it's a static generator I don't really think that's all that necessary.